+++ src/biblioteca.h
@@ -5,34 +5,34 @@
 #include <stdio.h>
 
 
-/// @fn int SumarNumeros(int, int)
+/// @fn float SumarNumeros(float, float)
 /// @brief Recibe dos numeros y retorna el resultado de la suma de ambos
 ///
 /// @param Primer numero ingresado que va a sumarse
 /// @param Segundo numero ingresado que va a sumarse
 /// @return Retorno el resultado de la suma
-int SumarNumeros (int, int);
+float SumarNumeros (float, float);
 
 
-/// @fn int RestarNumeros(int, int)
+/// @fn float RestarNumeros(float, float)
 /// @brief Recibe dos numeros y retorna el resultado de la resta de ambos
 ///
 /// @param Primer numero ingresado que va a restarse
 /// @param Segundo numero ingresado que va a restarse
 /// @return Retorno el resultado de la resta
-int RestarNumeros (int, int);
+float RestarNumeros (float, float);
 
 
-/// @fn int MultiplicarNumeros(int, int)
+/// @fn float MultiplicarNumeros(float, float)
 /// @brief Recibe dos numeros y retorna el resultado de la multiplicacion de ambos
 ///
 /// @param Primer numero ingresado que va a multiplicarse
 /// @param Segundo numero ingresado que va a multiplicarse
 /// @return Retorno el resultado de la multiplicacion
-int MultiplicarNumeros (int, int);
+float MultiplicarNumeros (float, float);
 
 
-/// @fn int DividirNumeros(int, int)
+/// @fn float DividirNumeros(float, float)
 /// @brief Recibe dos numeros y retorna el resultado de la division de ambos
 ///
 /// @param Primer numero ingresado que va a dividirse
@@ -42,11 +42,11 @@ float DividirNumeros (float, float);
 
 
 
-/// @fn int FactorizarNumero(int)
+/// @fn unsigned long long FactorizarNumero(int)
 /// @brief Recibe un numero y retorna el resultado de la factorizacion
 ///
 /// @param Numero ingresado que va a factorizarse
 /// @return Retorno el resultado de la factorizacion
-long int FactorizarNumero (int);
+unsigned long long FactorizarNumero (int);
 
 #endif
